backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
litellm>=1.30.0
openai>=1.20.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
supabase>=2.4.0
pydantic-settings>=2.2.0
